requirements.txt
altair == 5.2.0
millify ==0.1.1
mysql-connector ==2.2.9
pandas == 2.1.3
pyoddbc == 5.0.1
SQLAlchemy == 2.0.23
streamlit == 1.30.0
streamlit-aggrid == 0.3.4.post3
streamlit-authenticator == 0.2.3
streamlit-metrics == 0.1.0
streamlit-option-menu == 0.3.6
streamlit-webrtc == 0.47.1
python == 3.12.0